ngo_project/requirements.txt
Django==4.2.11
python-dotenv==1.0.0
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==0.7
gunicorn==21.2.0
Pillow>=8.0
